I think? it works can't test rn

+++ src/main/java/com/gregtechceu/gtceu/common/data/GTRecipeCapabilities.java
@@ -22,7 +22,7 @@ public class GTRecipeCapabilities {
     public final static RecipeCapability<FluidIngredient> FLUID = FluidRecipeCapability.CAP;
     public final static RecipeCapability<BlockState> BLOCK_STATE = BlockStateRecipeCapability.CAP;
     public final static RecipeCapability<Long> EU = EURecipeCapability.CAP;
-    public final static RecipeCapability<Integer> CWU = CWURecipeCapability.CAP;
+    public final static RecipeCapability<Long> CWU = CWURecipeCapability.CAP;
     public final static RecipeCapability<Float> SU = StressRecipeCapability.CAP;
 
     public static void init() {
